feat(rust): improve `RpcProxyService`
